Do not call deprecated h.singletonList
Jenkins 2.358 deprecated the `Functions.singletonList` method because
it is not normally needed in JEXL.  The `h.singletonList(it)` call in
this groovy code reports a null pointer exception in Jenkins 2.358 and
later.

When the `h.singletonList(it)` call is replaced with `it`, the null
pointer exception is not reported and the page loads.